chore(compass-editor): add codemirror autocompleter for stage editor COMPASS-6547 (#4163)
chore(compass-editor): add codemirror autocompleter for stage editor

+++ configs/mocha-config-compass/register/why-node-running.js
@@ -1,6 +1,9 @@
 require('why-is-node-running/include');
 exports.mochaHooks = {
   afterAll() {
+    if (process.argv.includes('--watch')) {
+      return;
+    }
     const timeout = setTimeout(() => {
       console.log(
         "if the process still running, run kill -SIGINFO %s to see what's keeping it",
